Fix : 지역 및 극장 리스트 수정 및 cinemaDetail로 정보가져오기

+++ src/main/java/com/sh/app/cinema/dto/CinemaDto.java
@@ -1,7 +1,5 @@
 package com.sh.app.cinema.dto;
 
-import com.sh.app.location.entity.Location;
-
 import lombok.Data;
 
 @Data
@@ -10,8 +8,8 @@ public class CinemaDto {
     private String region_cinema;
     private int theater_number;
     private String address;
-    private int location_lo;
-    private int location_la;
+    private Double location_lo;
+    private Double location_la;
     private String phone;
     private String location_name; // Location#location_name
 }
